modified:   calculations/forms.py 	new file:   calculations/management/__init__.py 	new file:   calculations/management/commands/__init__.py 	new file:   calculations/management/commands/update_salaries.py 	new file:   calculations/migrations/0005_alter_salary_options.py 	new file:   calculations/migrations/0006_alter_salary_base_end_time_and_more.py 	new file:   calculations/migrations/0007_alter_salary_day.py 	new file:   calculations/migrations/0008_alter_salary_day.py 	modified:   calculations/models.py 	deleted:    calculations/templates/calculations/display-salary-data.html 	new file:   calculations/templates/calculations/edit-salary.html 	modified:   calculations/templates/calculations/monthly_view.html 	modified:   calculations/urls.py 	modified:   calculations/views.py 	modified:   core/templates/core/index.html 	modified:   core/views.py

+++ calculations/models.py
@@ -11,10 +11,10 @@ class Salary (models.Model):
     recupero = models.BooleanField(default=False)
     daily_pay=  models.DecimalField(max_digits=10, decimal_places=2)
     day = models.DateField(unique=True)
-    base_start_time=  models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
-    base_end_time=  models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
-    extra_start_time=  models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
-    extra_end_time=  models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+    base_start_time=  models.IntegerField(null=True, blank=True)
+    base_end_time=   models.IntegerField(null=True, blank=True)
+    extra_start_time=   models.IntegerField(null=True, blank=True)
+    extra_end_time=   models.IntegerField(null=True, blank=True)
     
     def __str__(self):
          return f"Salary for {self.day}"
